+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - updated tests for set
 - updated tests for isExist
 - added test for randomElementFromSlice
+- added tests for urlsHaveSameDomain
 
 ## 0.2.1 - 2021.08.12